refactor(model_manager): relocate pipeline-run accessor to lib/shared
SourcePipelineRun/get_source_pipeline_run() reads generic pod-injected
identity env vars and isn't specific to model registration -- moved out
of model_manager/registry/ into the cross-cutting lib/shared/ package
(alongside the existing json_data/model_fuser/model_metadata utilities)
so other lib subpackages can use it without a model_manager dependency.
Pure relocation: no behavior change.
