+++ python/michelangelo/lib/shared/pipeline_run/__init__.py
@@ -1,0 +1,10 @@
+"""Pipeline-run identity, read back from the pod environment the worker injects."""
+
+# flake8: noqa:F401
+from .pipeline_run import (
+    SourcePipelineRun,
+    _reset_source_pipeline_run_cache,
+    get_source_pipeline_run,
+)
+
+__all__ = ["SourcePipelineRun", "get_source_pipeline_run"]
